+++ src/core/queue/simple/index.ts
@@ -11,17 +11,14 @@
  * @packageDocumentation
  */
 
-import Queue from 'core/queue/interface';
-import { Tasks } from 'core/queue/simple/interface';
-
+import Queue, { Tasks, CreateTasks } from 'core/queue/interface';
 export * from 'core/queue/interface';
-export * from 'core/queue/simple/interface';
 
 /**
  * Simple implementation of a queue data structure
  * @typeparam T - queue element
  */
-export default class SimpleQueue<T> implements Queue<T> {
+export default class SimpleQueue<T> extends Queue<T> {
 	/**
 	 * Type: list of tasks
 	 */
@@ -40,7 +37,13 @@ export default class SimpleQueue<T> implements Queue<T> {
 	/**
 	 * List of tasks
 	 */
-	protected tasks: this['Tasks'] = this.createTasks();
+	protected tasks: this['Tasks'];
+
+	/** @override */
+	constructor() {
+		super();
+		this.tasks = this.createTasks();
+	}
 
 	/** @inheritDoc */
 	push(task: T): number {
@@ -56,15 +59,13 @@ export default class SimpleQueue<T> implements Queue<T> {
 
 	/** @inheritDoc */
 	clear(): void {
-		if (this.tasks.length > 0) {
+		if (this.length > 0) {
 			this.tasks = this.createTasks();
 		}
 	}
 
 	/**
-	 * Return a new blank list of tasks
+	 * Returns a new blank list of tasks
 	 */
-	protected createTasks(): this['Tasks'] {
-		return [];
-	}
+	protected createTasks: CreateTasks<this['Tasks']> = () => [];
 }